Split legacy graph validation fixtures
